[REFACTOR] auth coordinator에 adaptor 폴더링
