+++ LionHeart-iOS/LionHeart-iOS/Scenes/Coordinator/CoordinatorImpl/AuthCoordinatorImpl.swift
@@ -7,45 +7,6 @@
 
 import UIKit
 
-protocol AuthCoordinator: Coordinator {
-    func showLoginViewController()
-    func showTabbarController()
-    func showOnboardingCompleteViewController(data: UserOnboardingModel)
-    func showTabbarOrOnboardingViewController(userState: UserState, kakaoToken: String?)
-    func pop()
-    func exitApplication()
-}
-
-typealias AuthNaviation = LoginNavigation & OnboardingNavigation & CompleteOnbardingNavigation
-
-final class AuthAdaptor: AuthNaviation {
-    
-    let coordinator: AuthCoordinator
-    init(coordinator: AuthCoordinator) {
-        self.coordinator = coordinator
-    }
-    
-    func checkUserIsVerified(userState: UserState, kakaoToken: String?) {
-        self.coordinator.showTabbarOrOnboardingViewController(userState: userState, kakaoToken: kakaoToken)
-    }
-    
-    func onboardingCompleted(data: UserOnboardingModel) {
-        self.coordinator.showOnboardingCompleteViewController(data: data)
-    }
-    
-    func startButtonTapped() {
-        self.coordinator.showTabbarController()
-    }
-    
-    func checkTokenIsExpired() {
-        self.coordinator.exitApplication()
-    }
-    
-    func backButtonTapped() {
-        self.coordinator.pop()
-    }
-}
-
 final class AuthCoordinatorImpl: AuthCoordinator {
     weak var parentCoordinator: Coordinator?
     private let factory: AuthFactory
