+++ LionHeart-iOS/LionHeart-iOS/Scenes/Coordinator/CoordinatorInterface/AuthCoordinator.swift
@@ -1,0 +1,17 @@
+//
+//  AuthCoordinator.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/17.
+//
+
+import Foundation
+
+protocol AuthCoordinator: Coordinator {
+    func showLoginViewController()
+    func showTabbarController()
+    func showOnboardingCompleteViewController(data: UserOnboardingModel)
+    func showTabbarOrOnboardingViewController(userState: UserState, kakaoToken: String?)
+    func pop()
+    func exitApplication()
+}
